pyproject.toml
[tool.poetry]
name = "todolist-backend"
version = "0.1.0"
description = ""
authors = ["namsu-k <nghokr1111@naver.com>"]
readme = "README.md"
packages = [{include = "todolist_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
django-environ = "^0.9.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
